document-intelligence/pyproject.toml
[project]
name = "sema4ai-docint"
version = "0.13.1"
description = "Sema4.ai Document Intelligence library"
readme = "README.md"
authors = [
    { name = "Sema4ai Developers", email = "engineering@sema4.ai" }
]
license = {file = "LICENSE"}
requires-python = ">=3.12,<4.0"
dependencies = [
    "sema4ai-data>=1.0.8,<2",
    "sema4ai-http-helper==2.1.0",
    "pydantic>=2.11,<3",
    "sema4ai.actions>=1.6.1,<2",
    "pandas>=2.2.3",
    "openpyxl>=3.1.5",
    "python-docx==0.8.11",
    "pymupdf>=1.24.10",
    "pillow>=10.0.0",
    "mindsdb-sql==0.21.0",
    "jsonschema>=4.0.0",
    "reductoai==0.9.0",
    "requests>=2.32.3",
    "scikit-learn>=1.7.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.1.0",
    "aiohttp>=3.10.0",
    "jinja2>=3.1.6",
]

[project.optional-dependencies]
memory = [
    "fastapi>=0.115.6,<0.116.0",
    "uvicorn>=0.34.0,<0.35.0",
    "python-multipart>=0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sema4ai_docint"]

[dependency-groups]
dev = [
    # Brings in memory dependencies for testing without requiring devs to use --all-extras
    "sema4ai-docint[memory]",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "psycopg2-binary>=2.9.6",
    "sema4ai-data>=1.0.2",
    # Upgrading to a newer version breaks on Josh's computer
    "testcontainers[postgres]==4.10.0",
    "deepdiff>=6.7.0",
    "sema4ai-data>=1.0.2",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "pyright>=1.1.390",
    "ruff>=0.7.2,<0.12.0",
    "pytest-repeat>=0.9.4",
    "twine>=6.1.0",
    "pytest-retry>=1.7.0",
    "pytest-xdist>=3.8.0",
    "psutil>=6.0.0",
    "pdoc>=14.0.0",
    "invoke>=2.2.0",
]

[tool.pytest.ini_options]
pythonpath = "."
markers = [
    "nl2sql_eval: mark test to run evals for nl to sql",
    "schema_eval: mark test to run schema generation evals",
]
filterwarnings = [
    # Silence PyMuPDF/SWIG deprecation warnings in test output
    "ignore:.*SwigPyPacked.*:DeprecationWarning",
    "ignore:.*SwigPyObject.*:DeprecationWarning",
    "ignore:.*swigvarlink.*:DeprecationWarning",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
    "@abstractmethod",
]
show_missing = true

[tool.ruff]
line-length = 100
show-fixes = true
respect-gitignore = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "PL",
    "N",
    "UP",
    "B",
    "A",
    "COM",
    "PT",
    "RUF",
    "W",
    "C90",
]
fixable = ["ALL"]
unfixable = []
ignore = ["COM812","PLC0415","N802"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["PLR2004"]
"tasks.py" = ["PT028"]

[tool.ruff.lint.pylint]
max-args = 10

[tool.ruff.lint.isort]
known-first-party = [
    "sema4ai_docint",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
include = ["sema4ai_docint"]
exclude = [
    "**/__pycache__",
    ".venv",
    "dist",
]
# Files to ignore during incremental type checking cleanup
# Remove files from this list as they are fixed
ignore = [
    "sema4ai_docint/__init__.py",
    "sema4ai_docint/agent_server_client/client.py",
    "sema4ai_docint/agent_server_client/docx_to_markdown.py",
    "sema4ai_docint/agent_server_client/images.py",
    "sema4ai_docint/agent_server_client/transport/http.py",
    "sema4ai_docint/agent_server_client/transport/memory.py",
    "sema4ai_docint/extraction/reducto/async_.py",
    "sema4ai_docint/extraction/reducto/client.py",
    "sema4ai_docint/models/data_model.py",
    "sema4ai_docint/models/document.py",
    "sema4ai_docint/models/document_layout.py",
    "sema4ai_docint/models/initialize.py",
    "sema4ai_docint/schemas/data_model.py",
    "sema4ai_docint/services/_document.py",
    "sema4ai_docint/services/_knowledge_base_service.py",
    "sema4ai_docint/services/_setup_kb.py",
    "sema4ai_docint/services/di_service.py",
    "sema4ai_docint/services/persistence/file/chat_file_accessor.py",
    "sema4ai_docint/validation/validate.py",
    "sema4ai_docint/views/schema.py",
]
pythonVersion = "3.12"
pythonPlatform = "All"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportUnknownParameterType = false
